make retrigger getter, remove hand_state and play_state getters/setters, and move hand_size, card_drawn, sound_played, discard_card, scored_card_index, hand_state and play_state to local variables to round.c
