+++ source/joker_effects.c
@@ -1,4 +1,5 @@
 #include "game.h"
+#include "game/round.h"
 #include "hand_analysis.h"
 #include "joker.h"
 #include "list.h"
